:truck: Rename model modules and verbose labels
This focuses less on the metadata aspect and more on the configuration
aspect, as these models are also used for runtime behaviour.
